Move some parsing code to a sub-module
