+++ generator/src/generator/namespace/mod.rs
@@ -6,7 +6,6 @@
 
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
-use std::convert::TryFrom;
 use std::rc::Rc;
 
 use xcbgen::defs as xcbdefs;
@@ -19,6 +18,7 @@ pub(super) mod helpers;
 mod header;
 mod request;
 mod serialize;
+mod parse;
 
 use helpers::{
     Caches, CaseInfo, DeducibleField, DeducibleFieldOp, DeducibleLengthFieldOp,
@@ -360,14 +360,15 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                         rust_field_type,
                     );
                     out.indented(|out| {
-                        self.emit_field_parse(
+                        parse::emit_field_parse(
+                            self,
                             field,
                             &rust_name,
                             "remaining",
                             FieldContainer::Other,
                             out,
                         );
-                        self.emit_field_post_parse(field, out);
+                        parse::emit_field_post_parse(field, out);
                         outln!(out, "let _ = remaining;");
                         outln!(out, "Ok({})", rust_field_name);
                     });
@@ -977,7 +978,8 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                     }
                     Self::emit_let_value_for_dynamic_align(fields, out);
                     for field in fields.iter() {
-                        self.emit_field_parse(
+                        parse::emit_field_parse(
+                            self,
                             field,
                             switch_prefix,
                             "remaining",
@@ -991,7 +993,7 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                             .map(|field_name| deducible_fields.contains_key(field_name))
                             .unwrap_or(false)
                         {
-                            self.emit_field_post_parse(field, out);
+                            parse::emit_field_post_parse(field, out);
                         }
                     }
                     let field_names = fields
@@ -1746,7 +1748,8 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                                     let case_fields = case.fields.borrow();
                                     Self::emit_let_value_for_dynamic_align(&*case_fields, out);
                                     for field in case_fields.iter() {
-                                        self.emit_field_parse(
+                                        parse::emit_field_parse(
+                                            self,
                                             field,
                                             name,
                                             "remaining",
@@ -1755,7 +1758,7 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                                         );
                                     }
                                     for field in case_fields.iter() {
-                                        self.emit_field_post_parse(field, out);
+                                        parse::emit_field_post_parse(field, out);
                                     }
                                     outln!(out, "outer_remaining = remaining;");
                                 }
@@ -1837,7 +1840,8 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                                     let case_fields = case.fields.borrow();
                                     Self::emit_let_value_for_dynamic_align(&*case_fields, out);
                                     for field in case_fields.iter() {
-                                        self.emit_field_parse(
+                                        parse::emit_field_parse(
+                                            self,
                                             field,
                                             name,
                                             "remaining",
@@ -1846,7 +1850,7 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
                                         );
                                     }
                                     for field in case_fields.iter() {
-                                        self.emit_field_post_parse(field, out);
+                                        parse::emit_field_post_parse(field, out);
                                     }
                                     outln!(out, "outer_remaining = remaining;");
                                 }
@@ -2254,327 +2258,6 @@ impl<'ns, 'c> NamespaceGenerator<'ns, 'c> {
         if has_dynamic_pad {
             outln!(out, "let value = remaining;");
         }
-    }
-
-    fn emit_field_parse(
-        &self,
-        field: &xcbdefs::FieldDef,
-        switch_prefix: &str,
-        from: &str,
-        container: FieldContainer,
-        out: &mut Output,
-    ) {
-        match field {
-            xcbdefs::FieldDef::Pad(pad_field) => {
-                let pad_size = match pad_field.kind {
-                    xcbdefs::PadKind::Bytes(size) => size.to_string(),
-                    xcbdefs::PadKind::Align(align) => {
-                        outln!(out, "// Align offset to multiple of {}", align);
-                        outln!(
-                            out,
-                            "let offset = remaining.as_ptr() as usize - value.as_ptr() as usize;",
-                        );
-                        outln!(
-                            out,
-                            "let misalignment = ({} - (offset % {})) % {};",
-                            align,
-                            align,
-                            align,
-                        );
-                        "misalignment".into()
-                    }
-                };
-                outln!(
-                    out,
-                    "let remaining = {from}.get({pad}..).ok_or(ParseError::InsufficientData)?;",
-                    from = from,
-                    pad = pad_size,
-                );
-            }
-            xcbdefs::FieldDef::Normal(normal_field) => {
-                let rust_field_name = to_rust_variable_name(&normal_field.name);
-                outln!(
-                    out,
-                    "let ({}, remaining) = {};",
-                    rust_field_name,
-                    self.emit_value_parse(&normal_field.type_, from),
-                );
-            }
-            xcbdefs::FieldDef::List(list_field) => {
-                let rust_field_name = to_rust_variable_name(&list_field.name);
-
-                if self.rust_value_type_is_u8(&list_field.element_type) {
-                    // List of `u8`, use simple parsing
-                    if let Some(list_length) = list_field.length() {
-                        outln!(
-                            out,
-                            "let ({}, remaining) = crate::x11_utils::parse_u8_list({}, {})?;",
-                            rust_field_name,
-                            from,
-                            list_length,
-                        );
-                        let ownership = if container == FieldContainer::Other {
-                            // Only force taking ownership for non-request types.
-                            ""
-                        } else {
-                            // Otherwise convert this into a reference to a fixed length array.
-                            // It's basically dumb luck that the largest list in the X11 protocol
-                            // and the largest fixed length array that Rust has a builtin conversion
-                            // for going from &[T] to &[T; N] both happen to be 32.
-                            assert!(list_length <= 32);
-                            "&"
-                        };
-                        outln!(
-                            out,
-                            "let {field} = <{ownership}[u8; \
-                             {length}]>::try_from({field}).unwrap();",
-                            field = rust_field_name,
-                            ownership = ownership,
-                            length = list_length,
-                        );
-                    } else if let Some(ref length_expr) = list_field.length_expr {
-                        outln!(
-                            out,
-                            "let ({}, remaining) = crate::x11_utils::parse_u8_list({}, \
-                             {}.try_to_usize()?)?;",
-                            rust_field_name,
-                            from,
-                            self.expr_to_str(
-                                length_expr,
-                                to_rust_variable_name,
-                                false,
-                                false,
-                                true,
-                            ),
-                        );
-                        // Only force taking ownership for non-request types.
-                        if container == FieldContainer::Other {
-                            outln!(
-                                out,
-                                "let {} = {}.to_vec();",
-                                rust_field_name,
-                                rust_field_name
-                            );
-                        }
-                    } else {
-                        // Only force taking ownership for non-request types.
-                        if container == FieldContainer::Other {
-                            outln!(out, "let {} = remaining.to_vec();", rust_field_name);
-                            outln!(out, "let remaining = &remaining[remaining.len()..];");
-                        } else {
-                            outln!(
-                                out,
-                                "let ({}, remaining) = remaining.split_at(remaining.len());",
-                                rust_field_name
-                            );
-                        }
-                    }
-                } else if self.can_use_simple_list_parsing(&list_field.element_type)
-                    && list_field.length_expr.is_some()
-                    && list_field.length().is_none()
-                {
-                    let rust_element_type =
-                        self.type_to_rust_type(list_field.element_type.type_.get_resolved());
-                    outln!(
-                        out,
-                        "let ({}, remaining) = crate::x11_utils::parse_list::<{}>(remaining, \
-                         {}.try_to_usize()?)?;",
-                        rust_field_name,
-                        rust_element_type,
-                        self.expr_to_str(
-                            list_field.length_expr.as_ref().unwrap(),
-                            to_rust_variable_name,
-                            false,
-                            false,
-                            true,
-                        ),
-                    );
-                } else if let Some(list_len) = list_field.length() {
-                    for i in 0..list_len {
-                        let tmp_name = format!("{}_{}", rust_field_name, i);
-                        outln!(
-                            out,
-                            "let ({}_{}, remaining) = {};",
-                            rust_field_name,
-                            i,
-                            self.emit_value_parse(&list_field.element_type, from),
-                        );
-                        self.emit_value_post_parse(&list_field.element_type, &tmp_name, out);
-                    }
-                    outln!(out, "let {} = [", rust_field_name);
-                    for i in 0..list_len {
-                        outln!(out.indent(), "{}_{},", rust_field_name, i);
-                    }
-                    outln!(out, "];");
-                } else {
-                    outln!(out, "let mut remaining = {};", from);
-                    if let Some(ref length_expr) = list_field.length_expr {
-                        outln!(
-                            out,
-                            "let list_length = {}.try_to_usize()?;",
-                            self.expr_to_str(
-                                length_expr,
-                                to_rust_variable_name,
-                                false,
-                                false,
-                                false,
-                            ),
-                        );
-                        outln!(
-                            out,
-                            "let mut {} = Vec::with_capacity(list_length);",
-                            rust_field_name
-                        );
-                        outln!(out, "for _ in 0..list_length {{");
-                    } else {
-                        outln!(out, "// Length is 'everything left in the input'");
-                        outln!(out, "let mut {} = Vec::new();", rust_field_name);
-                        outln!(out, "while !remaining.is_empty() {{");
-                    }
-                    out.indented(|out| {
-                        outln!(
-                            out,
-                            "let (v, new_remaining) = {};",
-                            self.emit_value_parse(&list_field.element_type, from),
-                        );
-                        self.emit_value_post_parse(&list_field.element_type, "v", out);
-                        outln!(out, "remaining = new_remaining;");
-                        outln!(out, "{}.push(v);", rust_field_name);
-                    });
-                    outln!(out, "}}");
-                }
-            }
-            xcbdefs::FieldDef::Switch(switch_field) => {
-                let rust_field_name = to_rust_variable_name(&switch_field.name);
-                let switch_struct_name = if let FieldContainer::Request(request_name) = container {
-                    format!("{}Aux", request_name)
-                } else {
-                    format!("{}{}", switch_prefix, to_rust_type_name(&switch_field.name))
-                };
-                let mut parse_params = vec![String::from("remaining")];
-                for ext_param in switch_field.external_params.borrow().iter() {
-                    parse_params.push(to_rust_variable_name(&ext_param.name));
-                }
-                outln!(
-                    out,
-                    "let ({}, remaining) = {}::try_parse({})?;",
-                    rust_field_name,
-                    switch_struct_name,
-                    parse_params.join(", "),
-                );
-            }
-            xcbdefs::FieldDef::Fd(fd_field) => {
-                let rust_field_name = to_rust_variable_name(&fd_field.name);
-                outln!(
-                    out,
-                    "if fds.is_empty() {{ return Err(ParseError::MissingFileDescriptors) }}"
-                );
-                outln!(out, "let {} = fds.remove(0);", rust_field_name);
-            }
-            xcbdefs::FieldDef::FdList(fd_list_field) => {
-                let rust_field_name = to_rust_variable_name(&fd_list_field.name);
-
-                outln!(
-                    out,
-                    "let fds_len = {}.try_to_usize()?;",
-                    self.expr_to_str(
-                        &fd_list_field.length_expr,
-                        to_rust_variable_name,
-                        false,
-                        false,
-                        false,
-                    ),
-                );
-                outln!(
-                    out,
-                    "if fds.len() < fds_len {{ return Err(ParseError::MissingFileDescriptors) }}",
-                );
-                outln!(out, "let mut {} = fds.split_off(fds_len);", rust_field_name);
-                outln!(out, "std::mem::swap(fds, &mut {});", rust_field_name);
-            }
-            xcbdefs::FieldDef::Expr(expr_ref) => {
-                match expr_ref.expr {
-                    xcbdefs::Expression::Value(_) => {
-                        // Treat this as a normal field, that we'll validate in
-                        // emit_field_post_parse.
-                        let rust_field_name = to_rust_variable_name(&expr_ref.name);
-                        outln!(
-                            out,
-                            "let ({}, remaining) = {};",
-                            rust_field_name,
-                            self.emit_value_parse(&expr_ref.type_, from),
-                        );
-                    }
-                    _ => {
-                        // The only non-constant expr field we ever need to parse is
-                        // odd_length, so assert that we have that and then just write
-                        // some one-off code for it.
-                        assert_eq!(expr_ref.name, "odd_length");
-                        outln!(
-                            out,
-                            "let (odd_length, remaining) = bool::try_parse(remaining)?;"
-                        );
-                        // And there is a matching special case to use this in request parsing.
-                    }
-                }
-            }
-            xcbdefs::FieldDef::VirtualLen(_) => {}
-        }
-    }
-
-    fn emit_field_post_parse(&self, field: &xcbdefs::FieldDef, out: &mut Output) {
-        match field {
-            xcbdefs::FieldDef::Normal(normal_field) => {
-                let rust_field_name = to_rust_variable_name(&normal_field.name);
-                self.emit_value_post_parse(&normal_field.type_, &rust_field_name, out);
-            }
-            xcbdefs::FieldDef::Expr(xcbdefs::ExprField {
-                name,
-                expr: xcbdefs::Expression::Value(v),
-                ..
-            }) => {
-                assert!(u8::try_from(*v).is_ok());
-                let rust_field_name = to_rust_variable_name(name);
-                outln!(out, "if {} != {} {{", rust_field_name, v);
-                outln!(out.indent(), "return Err(ParseError::InvalidValue);");
-                outln!(out, "}}");
-            }
-            _ => (),
-        }
-    }
-
-    fn emit_value_parse(&self, type_: &xcbdefs::FieldValueType, from: &str) -> String {
-        let type_type = type_.type_.get_resolved();
-        let rust_type = self.type_to_rust_type(type_type);
-        let params = self.get_type_parse_params(type_type, from);
-        format!("{}::try_parse({})?", rust_type, params.join(", "))
-    }
-
-    fn emit_value_post_parse(
-        &self,
-        type_: &xcbdefs::FieldValueType,
-        var_name: &str,
-        out: &mut Output,
-    ) {
-        if let xcbdefs::FieldValueSet::Enum(_) = type_.value_set {
-            // Handle turning things into enum instances.
-            outln!(out, "let {var} = {var}.into();", var = var_name);
-        }
-    }
-
-    fn needs_post_parse(&self, type_: &xcbdefs::FieldValueType) -> bool {
-        if let xcbdefs::FieldValueSet::Enum(_) = type_.value_set {
-            true
-        } else {
-            false
-        }
-    }
-
-    fn can_use_simple_list_parsing(&self, type_: &xcbdefs::FieldValueType) -> bool {
-        self.get_type_parse_params(&type_.type_.get_resolved(), "")
-            .len()
-            == 1
-            && !self.needs_post_parse(type_)
     }
 
     fn emit_calc_deducible_field(
